Rename to typedecl
I realized that enums had extra connotations than just "all the
enumerable values of this type" and that it lead to more confusion than
not. "typedecl" is very clearly about type declarations in a very Go
naming convention way, so it seems like a decent compromise towards
clarity as I'm preparing to tag v1.0. :)